refactor: move code of menu edit task to separate file
